requirements.txt
# algo_demo_india — pinned dependencies
flask==3.0.3
flask-socketio==5.3.7
flask-cors==4.0.1
eventlet==0.36.1
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.2
scikit-learn==1.5.1
xgboost==2.1.0
joblib==1.4.2
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.2
